Feat : 유저 삭제시 userId도 받아주기 체크하기위해서

+++ src/main/java/com/developer/wiki/oauth/controller/UserController.java
@@ -1,16 +1,14 @@
 package com.developer.wiki.oauth.controller;
 
 
+import com.developer.wiki.common.exception.BadRequestException;
 import com.developer.wiki.oauth.User;
 import com.developer.wiki.oauth.dto.UserResponseDto;
 import com.developer.wiki.oauth.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
 @RestController
@@ -23,8 +21,9 @@ public class UserController {
         return ResponseEntity.ok().body(currentUser.toDto());
     }
 
-    @DeleteMapping
-    public ResponseEntity<String> deleteUSer(@AuthenticationPrincipal User currentUser){
+    @DeleteMapping("/{userId}")
+    public ResponseEntity<String> deleteUSer(@AuthenticationPrincipal User currentUser, @PathVariable(name = "userId")Long userid){
+        if(!currentUser.getId().equals(userid)) throw new BadRequestException("Not Match Userid");
         userService.deleteUser(currentUser.getId());
         return ResponseEntity.ok("delete");
     }
